+++ eval/src/exec/mod.rs
@@ -51,22 +51,25 @@ pub use crate::compiler::CompilerExt;
 /// # use hashbrown::HashSet;
 ///
 /// # fn main() -> anyhow::Result<()> {
-/// let module = Untyped::<F32Grammar>::parse_statements("fold(xs, -INFINITY, max)")?;
+/// let module = Untyped::<F32Grammar>::parse_statements(
+///     "Array::fold(xs, -INFINITY, max)",
+/// )?;
 /// let module = ExecutableModule::new("test", &module)?;
 ///
 /// let mut env = Environment::new();
-/// env.extend(env::Prelude::vars());
 /// env.extend(env::Comparisons::vars());
+/// env.extend(env::Prelude::prototypes());
+/// env.insert_prototypes_as_vars();
 /// env.insert("INFINITY", Value::Prim(f32::INFINITY)).insert("xs", Value::void());
 ///
 /// // With the original imports, the returned value is `-INFINITY`.
 /// assert_eq!(module.with_env(&env)?.run()?, Value::Prim(f32::NEG_INFINITY));
 ///
 /// // Imports can be changed. Let's check that `xs` is indeed an import.
-/// assert!(module.is_import("fold"));
+/// assert!(module.is_import("xs"));
 /// // It's possible to iterate over imports, too.
 /// let imports = module.import_names().collect::<HashSet<_>>();
-/// assert!(imports.is_superset(&HashSet::from_iter(vec!["max", "fold"])));
+/// assert!(imports.is_superset(&HashSet::from_iter(vec!["max", "Array"])));
 /// # drop(imports); // necessary to please the borrow checker
 ///
 /// // Change the `xs` import and run the module again.
@@ -103,7 +106,7 @@ pub use crate::compiler::CompilerExt;
 ///
 /// ## Behavior on errors
 ///
-/// `run_in_env` modifies the environment even if an error occurs during execution:
+/// [`Self::with_mutable_env()`] modifies the environment even if an error occurs during execution:
 ///
 /// ```
 /// # use arithmetic_parser::grammars::{F32Grammar, Parse, Untyped};