+++ eval/src/values/object.rs
@@ -17,7 +17,7 @@ use arithmetic_parser::StripCode;
 
 /// Object with zero or more named fields.
 ///
-/// An object functions similarly to a [`HashMap`] with [`String`] keys and [`Value`]
+/// An object functions similarly to a `HashMap` with [`String`] keys and [`Value`]
 /// values. It allows iteration over name-value pairs, random access by field name,
 /// inserting / removing fields etc.
 ///
@@ -224,7 +224,7 @@ where
     }
 }
 
-/// Prototype of an [`Object`] or a [`Tuple`](crate::Tuple).
+/// Prototype of an [`Value`].
 ///
 /// A prototype is quite similar to an [`Object`]; it is a collection of named [`Value`]s.
 /// Prototype fields are used for method lookup similar to JavaScript; if an value has a prototype,
@@ -241,9 +241,9 @@ where
 /// to object fields.
 ///
 /// A prototype can be converted to a [`Function`], and prototypes defined in the script code
-/// *are* callable. Such a function associates the prototype with the provided value
-/// (an object or a tuple). For values without an associated prototype, method resolution
-/// is performed using [`StandardPrototypes`], which can be set for an [`Environment`].
+/// are callable. Such a function associates the prototype with the provided value,
+/// which must be an object. For values without an associated prototype, method resolution
+/// is performed using prototypes for standard types, which can be set in an [`Environment`].
 ///
 /// Prototypes can be defined both in the host code, and in scripts via [`CreatePrototype`].
 ///
@@ -273,7 +273,7 @@ where
 /// let program = r#"
 ///     pt.len() == 2 &&
 ///     Object(#{ foo: 1 }).len() == 1 &&
-///     (Object.len)(Object.EMPTY) == 0
+///     Object::len(Object.EMPTY) == 0
 /// "#;
 /// let program = Untyped::<F32Grammar>::parse_statements(program)?;
 /// let module = ExecutableModule::new("test_proto", &program)?;
@@ -347,19 +347,13 @@ impl<'a, T> Prototype<'a, T> {
         ctx.check_args_count(&args, 1)?;
         let mut arg = args.pop().unwrap();
 
-        match &mut arg.extra {
-            Value::Object(object) => {
-                object.set_prototype(self.clone());
-            }
-            Value::Tuple(tuple) => {
-                tuple.set_prototype(self.clone());
-            }
-            _ => {
-                let err = ErrorKind::native("Function argument must be an object or tuple");
-                return Err(ctx
-                    .call_site_error(err)
-                    .with_span(&arg, AuxErrorInfo::InvalidArg));
-            }
+        if let Value::Object(object) = &mut arg.extra {
+            object.set_prototype(self.clone());
+        } else {
+            let err = ErrorKind::native("Function argument must be an object");
+            return Err(ctx
+                .call_site_error(err)
+                .with_span(&arg, AuxErrorInfo::InvalidArg));
         }
         Ok(arg.extra)
     }
@@ -458,7 +452,7 @@ impl<T: Clone> StandardPrototypes<T> {
 ///
 /// # Examples
 ///
-/// See [`Environment`] docs for an example of usage.
+/// See [`Environment`](crate::Environment) docs for an example of usage.
 #[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
 pub struct PrototypeField {
     ty: ValueType,